Move Geo spot button onto map
