+++ public/stylesheet/style.css
@@ -472,3 +472,12 @@ form ul li {
 .directions-btn-exit:hover {
   background-color: #e2e2e2!important;
 }
+
+/* Map Buttons =================================== */
+.create-btn-geo {
+  margin: 15px;
+  width: 25%;
+  padding: 10px;
+  pointer-events: all;
+}
+
